+++ bin/migrate-oats-data/applications/__init__.py
@@ -9,4 +9,4 @@ from .migrate_application import (
     init_applications,
 )
 
-from .set_hide_from_portal_on_application import set_application_visibility
+from .set_application_visibility import set_application_visibility
