+++ bin/migrate-oats-data/applications/set_application_visibility.py
