+++ bin/migrate-oats-data/noi/__init__.py
@@ -1,3 +1,3 @@
 from .notice_of_intent_init import init_notice_of_intents, clean_notice_of_intents
 from .notice_of_intent_migration import process_notice_of_intent
-from .set_hide_from_portal_on_noi import set_notice_of_intent_visibility
+from .set_notice_of_intent_visibility import set_notice_of_intent_visibility
